Refactored Countable Streams to use long instead of int for its size management.

+++ sample/src/main/java/com/taig/communicator/sample/io/Events.java
@@ -54,7 +54,7 @@ public class Events extends Interaction
 			}
 
 			@Override
-			protected void onReceive( int current, int total )
+			protected void onReceive( long current, long total )
 			{
 				getTextView().setText( "RECEIVE (" + current + " / " + total + " kB)" );
 			}
